fix: invert started_at, ended_at columns nullability
The columns were assigned the incorrect nullability values upon creation. It did not allow for a shift to be added without an end time.

feat: add function to get the current spot by guard id

Added a helper endpoint to search for the current spot of a guard via table joins.